Renamed class files to pascal case
