+++ lib/HandlebarsCompiler.ts
