+++ lib/types/plugin-options/index.ts
@@ -1,4 +1,4 @@
-import { CompileOptions } from "../../handlebars-compiler";
+import { CompileOptions } from "../../HandlebarsCompiler";
 import { IPluginOptions } from "./params";
 
 export interface HandlebarsPluginOptions extends CompileOptions, IPluginOptions  {
